refactor:优化附件管理中对上传驱动的使用

+++ app/common/model/Attachment.php
@@ -44,7 +44,8 @@ class Attachment extends Base
 
     public function getFullUrlAttr($value, $row): string
     {
-        return self::$upload->getDriver($row['storage'])->url($row['url']);
+        $driver = self::$upload->getDriver($row['storage'], false);
+        return $driver ? $driver->url($row['url']) : full_url($row['url']);
     }
 
     /**
@@ -59,7 +60,8 @@ class Attachment extends Base
             ['storage', '=', $model->storage],
         ])->find();
         if ($repeat) {
-            if (!self::$upload->getDriver($repeat->storage)->exists($repeat->url)) {
+            $driver = self::$upload->getDriver($repeat->storage, false);
+            if ($driver && !$driver->exists($repeat->url)) {
                 $repeat->delete();
                 return true;
             } else {
@@ -93,8 +95,8 @@ class Attachment extends Base
     {
         Event::trigger('AttachmentDel', $model);
 
-        $driver = self::$upload->getDriver($model->storage);
-        if ($driver->exists($model->url)) {
+        $driver = self::$upload->getDriver($model->storage, false);
+        if ($driver && $driver->exists($model->url)) {
             $driver->delete($model->url);
         }
     }
